SQL file handling: Improve object name parser
Skeema parses the beginning of some SQL statements, to extract an object name
(e.g. schema name and/or table name) and associate that statement with its
corresponding live database object. However, the previous iteration of the
parser had several deficiencies, and was not easily extendable to support
additional types of objects.

This commit replaces the homegrown parser with one based on external dep
github.com/alecthomas/participle (MIT license), which is now vendored.